+++ worker/uniter/runner/context/mocks/payload_mock.go
@@ -8,7 +8,7 @@ import (
 	reflect "reflect"
 
 	gomock "github.com/golang/mock/gomock"
-	payload "github.com/juju/juju/payload"
+	payload "github.com/juju/juju/core/payload"
 )
 
 // MockPayloadAPIClient is a mock of PayloadAPIClient interface.
