+++ worker/uniter/runner/jujuc/context.go
@@ -17,8 +17,8 @@ import (
 	"github.com/juju/juju/core/application"
 	"github.com/juju/juju/core/life"
 	"github.com/juju/juju/core/network"
+	"github.com/juju/juju/core/payload"
 	"github.com/juju/juju/core/relation"
-	"github.com/juju/juju/payload"
 	"github.com/juju/juju/rpc/params"
 	"github.com/juju/juju/storage"
 )
